[SYSTEMML-1986] New nary cbind/rbind (compiler/runtime for cp/spark)
This patch generalizes the existing cbind and rbind from binary to nary
operations, which avoids unnecessary materialized intermediates and
simplifies its usage. Depending on the number of inputs, we now either
compile a BinaryOp or NaryOp which both however share a common block
runtime. For distributed operations, we support a new physical append
operator that relies on a simple "shift-pad-union-merge" approach and
thus, avoids unnecessary repeated shuffling. 

Additionally, this patch also adds the related tests for rbind and cbind
with three inputs and all combinations of sparse and dense as well as
modifies some existing scripts to exploit the new capabilities.