[SYSTEMML-1989] Performance external UDF invocation (loading, parsing)
This patch improves the performance of external (i.e., java) UDF
functions by moving the class loading and input/output parsing out of
the critical path into the instruction setup. Furthermore, this includes
a number of cleanups and updates of the UDF framework.

On a test scenario of a pass-through external UDF with three inputs and
output parameters, this patch improved performance from 75K/s to 267K/s,
which good considering the rate of 520K/s for dml-bodied functions.