[SYSTEMML-1988] Fix cloning of prepared scripts, rework CP runtime
This patch fixes the shallow cloning of prepared scripts by (1) better
handling of external functions (which require a deep copy), and (2) a
complete rework of our CP runtime instructions and operators to make
them read-only in order to avoid side effects. As a positive side effect
of this rework, the runtime is now much easier to debug as side effects
are impossible due to final members of all instructions and operators.