[SYSTEMML-1984] Fix robustness JMLC config mgmt, read from input streams
This patch fixes special cases of JMLC configuration management,
specifically, if a connection is created in one thread, but it is used
in other threads to read data from input streams. Due to lost
thread-local configurations, this API call failed trying to read via a
multi-threaded reader. We now ensure that reads have the proper local
configurations and also avoid unnecessary exceptions by falling back to
sequential readers if necessary.