[SYSTEMML-2007] New order builtin w/ multiple order-by columns
This patch generalizes the compiler and runtime of the existing order
builtin function to support multiple order-by columns. So far we only
support a single scalar column parameter - now users can either provide
a scalar or matrix argument. This avoids emulating this functionality
via multiple order calls in reverse order of order-by columns, which
works due to stable sort. 

This patch modifies the compiler and CP runtime, without degrading
performance for the case of a single order-by column. The spark runtime
will follow in a subsequent patch, as it requires an second sort runtime
to avoid unnecessary overhead in case of a single order-by column.

On a scenario of a 1M x 100 dense matrix, with varying number of
distinct values, and 3 order-by columns, this patch improved performance
(compared to the emulation via 3 order calls) as follows:

1M x 10, distinct=1000: 2,271ms -> 940ms
1M x 10, distinct=100: 2,125ms -> 966ms
1M x 10, distinct=10: 2.057ms -> 958ms