[SYSTEMML-2009] Performance multi-threaded JMLC scoring (contention)
This patch makes a major performance improvement to multi-threaded JMLC
scoring and script execution. For scripts with many small intermediates,
concurrent JMLC prepared scripts were contended on sequence ID
generators for unique matrix IDs, which is unnecessary because these IDs
are only used in the bufferpool which is disabled through JMLC. On an
end-to-end application with above characteristics, this changed improved
performance from 487s to 94s (in a test environment w/ 24 vcores).