[SYSTEMML-2010] Improved rewrite for merging statement blocks
So far we did not merging basic blocks that contain functions with any
subsequent blocks because function directly bind to output variables.
However, this is too conservative because blocks with non-conflicting
inputs and outputs can still be merged, which creates more opportunities
for rewrites and common subexpression elimination. This patch
generalizes the existing rewrite for merging sequences of statement
blocks accordingly. 

Furthermore, this patch also includes additional tests and a fix for a
potential result correctness issue due to the invalid merge of statement
blocks, where the second contains a function and there is an output
conflict.