[SYSTEMML-2013] Fix robustness spark nary cbind and rpc configuration
This patch addresses a recent perftest issue with stratstats on the 80GB
scenario. First, we now automatically configure (depending on the spark
version) either 'spark.akka.frameSize' or 'spark.rpc.message.maxSize' to
a more robust value of 512MB as we have done manually in the past.
Second, this patch improves the robustness of the recently added nary
cbind to use a preferred number of output partitions for the final
shuffle, which avoids large partitions in case of many inputs.