[SYSTEMML-2017] Fix missing bufferpool cleanup of frame intermediates
So far any rmvar, cpvar, mvvar instructions only trigger the cleanup of
buffer pool in-memory objects, evicted files and hdfs files for matrices
but not frames. This creates a leak of objects and files that causes
severe performance issues if frame intermediates are creates in
iterative algorithms. This patch now generalizes all related code paths
to CacheableData which is the super class of MatrixObjects and
FrameObjects.