[SYSTEMML-2021] Fix robustness codegen optimizer (opening heuristic)
This patch fixes the robustness of the codegen plan enumeration, which
got stuck for the perftest stratstats due to a large fusion partition
with 38 interesting points and very specific cost characteristics that
rendered the open heuristic fuse-all ineffective. We now run both
heuristics (fused-all and fuse-no-redundancy) as a consolidated opening
heuristics to quickly obtain a good lower bound. Furthermore, we also
avoid enumerating large partitions (w/ >20 interesting points) if the
opening heuristic already achieved costs that are in an epsilon
environment  (1%) of the minimal static costs of this partition.
Together, these modifications greatly improve the robustness of the
codegen optimizer without missing any meaningful plans.