[SYSTEMML-2022] Fix codegen blocksize handling for spark row ops
This patch addresses perftest failures of MLogreg sparse with intercept
2, where we incorrectly chosen a row template despite an output size of
1001. This was because (1) the memory estimates did not correctly
reflect the handling of sparse outputs in row templates, and (2) we only
checked valid input dimensions (in this case 1000) but not the output
dimensions (in this case 1001 after cbind). Furthermore, this patch also
improves the memory estimates for cell and outer templates that are able
to directly output sparse outputs without conversion.