[SYSTEMML-2028] Fix codegen invalid fallback from row to cell ops
This patch fixes issues of perftest Mlogreg 10M x 1K, icpt1 with codegen
due to unsupported cbind operations in cell templates. The root cause
was a violated block size constraint which led to an invalid fallback
from row to cell operations although these are not supported for cell
templates. With this patch, the fallback happens in an operator aware
manner, removing the entry in case of invalid fallbacks.