[SYSTEMML-2030] Improved transitive exec type selection spark mm
This patch improves the transitive execution type selection of spark
matrix multiply for cases where the input is produced by a spark
transpose operation and the transpose-mm rewrite is not applicable due
to memory constraints. On the perftest L2SVM 800GB sparse icpt 1, this
patch improved performance from 2,420s to 223s.