[SYSTEMML-1421] Remove unnecessary checks for Java 7 and below
We dropped the support for Java 6 and 7 a while ago and meanwhile, we
leverage Java 8 features such as lambda expressions expensively, which
means that SystemML will not run with Java 7 or below. Accordingly, this
patch removes unnecessary checks for Java versions < 8 (which had severe
scalability issues for multi-threaded string-double parsing - due to
internal static synchronized caches - as used in text and csv readers).