[SYSTEMML-2038] Memory-efficiency dense-sparse matrix conversion (CSR)
This patch reworks the dense-sparse matrix block conversion for improved
memory efficiency. This operation primitive knows the number of
non-zeros upfront and is a simply sequential operation. In order to
improve memory-efficiency of sparse intermediates, we now output the
target in CSR instead of MCSR format.

Furthermore, this also patch cleans up unnecessary (and misleading)
statistics maintenance in these conversion primitives, because these are
never called in an operation-specific manner (except for left indexing
operations).