[SYSTEMML-2039] Rework instruction generation for memory-efficiency
This patch makes a fundamental change to the instruction generation for
CP and SPARK execution types. So far, we sorted lops by level and ID
leading to a breadth-first instruction scheduling; for large DAGs, this
led to an unnecessarily large number of live intermediates, causing OOMs
in JMLC, unnecessary evictions in normal execution, and poor data
locality. The reason for breadth-first instruction scheduling, was to
ensure that transient writes do not overwrite the transient reads and in
order to support vertical piggybacking for MR jobs.

The new approach for CP and SPARK uses a simple two-level partitioning,
where the lops of a DAG are partitioning into leafs/intermediates and
outputs. All leafs/intermediates are scheduled in a depth-first manner,
followed by all outputs. There is room for even better scheduling (after
the 1.0 release) but this patch aims for minimal changes to the core
instruction generation, while removing the main bottleneck.

For a complex JMLC application, this patch reduced the maximum memory
requirements by more than 3x.