[SYSTEMML-2052] Fix ultra-sparse dataset-matrix conversion
This patch fixes special cases of converting ultra-sparse datasets to
binary block matrices. Specifically, there was an issue of
pre-allocating the MCSR sparse rows, which resulted in
index-out-of-bounds exceptions on preallocated rows of size zero. We now
correctly determine the nnz per block range and made the MCSR format
more robust to handling resizing rows with initial size zero.

Furthermore, this also includes additional tests for ultra-sparse
datasets and a performance improvement for counting the nnz for index
ranges of sparse dataset vectors.