[SYSTEMML-2046,2049] Enable large dense blocks, incl rand, write, alloc
This patch enables the runtime allocation of large dense blocks (but
unless forced to single node, the compiler falls back to distributed
operations) along with proper exception handling to indicate operations
that only work with a single dense blocks. This allows now end-to-end
experiments without internal changes.

Furthermore, this adds support for large dense blocks in rand operations
and local evictions. Additionally, this also includes a fix for the
allocation of large dense blocks which physical data size computed
incorrectly so far.