[SYSTEMML-2058] Fix sparse-unsafe, sparse-dense unary block operations
This patch fixes a special case of sparse-unsafe unary operations for
sparse inputs and dense outputs. These operations reset the output to
the constant value op(0) and subsequently redirect to sparseSafe
operations which in turn use append to add computed values. However,
this append rejects computed 0s (as is the case for NOT), leading to
incorrect results. We now directly update the dense target and also
ensure that the reset uses the double instead of long function because
the latter might not be implemented for all function objects.