[SYSTEMML-2062] Performance ultra-sparse transpose (adaptive blocking)
This patch improves the performance of sparse-sparse transpose
operations for ultra-sparse matrices. So far, we simply uses a blocksize
of 128 x 128 (equivalent to dense) for cache-conscious operations. For
ultra-sparse matrices, this creates a lot of overhead for maintaining
the block boundaries. We now use an adaptive (sparsity-aware) block size
that is derived from the input sparsity. 

On scenarios of executing 100 transpose operations, the total runtime
improved as follows:
a) 100K x 100K, sp=1e-4: 12.2s --> 5.2s.
b) 1M x 1M, sp=1e-6:     >25min --> 24.4s.