[SYSTEMML-2048] Compiler integration large dense blocks (no fallbacks)
After modifying all runtime operations to work with large dense blocks,
this patch now integrate these blocks into the compiler. So far, we
compiled distributed operations if an intermediate was potentially dense
and larger than 16GB in dense representation. With the new block
runtime, this fallback to distributed operations is unnecessary.