[SYSTEMML-2069] Lock-free maintenance of heavy hitter statistics
This patch addresses issues of thread contention on the maintenance of
heavy hitter statistics. In multi-threaded applications with many small
operations, the synchronized maintenance of these heavy hitters can turn
into a bottleneck. Example scenarios are JMLC scoring with multiple JMLC
streams per node and local parfor scripts.

We simply replace the synchronized maintenance with the use of a
concurrent hash map as well as a consolidated statistics objects for
times and counts, which rely on the scalable LongAdder primitive.

On an end-to-end JMLC scoring scenario, this patch improved the total
runtime from 261s to 117s (with statistics enabled), which is good
compared to the baseline of 101s (without statistics) because the time
measurement itself is also expensive. Similarly, end-to-end algorithms
such as decision tree and random forest will see similar benefits for
the case of enabled statistics.