[SYSTEMML-2050] Performance ifelse w/ constant expression (shallow copy)
This patch significantly improves the performance of ifelse ternary
operations for the case of scalar expression predicates or constant
matrix expressions. Specifically, we now use a conditional shallow copy
of the input if it's safe to do so.

On a scenario of 20 iterations, inputs of size 10K x 10K (dense), and
ifelse(TRUE, X, Y), the total runtime improved from 23.3s to 6ms
(~4000x). Similarly, for ifelse(TRUE, TRUE, Y), the total runtime
reduced to 5.7s (~4x).