[SYSTEMML-2091] Performance sparse-sparse reshape (shallow for CSR)
This patch improves the performance of sparse-sparse reshape operations
for the special case of N:1 row mappings and inputs in CSR format.
Specifically, we now allocate the output in the memory-efficient CSR
representation as well which allows for a shallow copy of the values
array. The row pointers and column indexes are, however, reconstructed.

On a scenario of 10 iterations of reshaping a 100K x 10K matrix with
sparsity 0.1 to a 10K x 100K matrix, this patch improved the total
runtime from 16.2s to 2.6s