[SYSTEMML-2093] Parser and language support for new += operator
This patch adds parser and language support for the new += accumulation
assignment operator. The semantics of A += B are equivalent to A = A +
B, but this operator will receive special treatment for parfor result
variables (with sum as result aggregation function). This patch also
modifies a number of existing algorithms accordingly to integrate this
operator into our testsuite.