[SYSTEMML-2099] Performance sparse result merge accum w/o compare
This patch improves the performance of dense-sparse binary in-place
operations for plus and minus (as used for accumulating result merge
without compare). On a scenario of merging 12 sparse results of size 10K
x 10K, sp=0.2, this patch improved performance from 53.3s to 4.9s. 

Furthermore, this also includes a minor fix to disable parfor update in
place for accumulating result variables.