[SYSTEMML-1437] Factorization Machines: Core Layer Module
This implements the core model of a factorization machine (Rendle 2010)
as a layer in the `nn` library.  The `forward` function accepts the
input data `X` and the parameters `w0`, `W`, and `V`, and returns the FM
output `y`, which is a vector containing a single output for each
example. The `backward` function accepts the upstream gradients w.r.t.
`y` (i.e., `dloss/dy`, the gradient of the loss function w.r.t. `y`),
which is the same shape as `y`, as well as `X`, `w0`, `W`, and `V`, and
returns the gradients of the loss w.r.t. `X` and the parameters, i.e.,
`dX`, `dw0`, `dW`, and `dV`. The `init` function accepts the number of
features `d` and the factorization dimensionality `k`, and returns
initialized `w0`, `W`, and `V` values.

In the future, scripts can be written that use this `fm` layer along
with other layer functions to implement factorization machines for
regression, classification, etc.

Closes #696.