[SYSTEMML-2108] Performance CP ternary +* and -* operations
Since the introduction of the general ternary operation framework for
ifelse (which also subsumed the specific +* and -* operations), the +*
and -* operations showed non-negligible overhead, especially for
sparse-dense combinations. Hence, this patch adds a special case for
matrix-scalar-matrix and matrix-matrix-scalar operations that routes
these operations to the binary operation framework.

On lenet over mnist, +* and -* consumed 28% execution time - this patch
then reduced the runtime of these operations by more than 2x.