[SYSTEMML-2120] New plan cache for cost-based codegen optimizer
So far the codegen framework used a plan cache to reused generated and
compiled operators. Since this operator plan cache determined operator
equivalence based on hashing cplans, the optimizer still had to
repeatedly enumerate and cost plans (which cover multiple cplans). This
patch introduces an additional plan cache in the cost-based optimizer to
reduce the overhead of repeated optimization with equivalent inputs.

In detail, we compute plan partition signatures and used these to match
equivalent optimization problems which allows to reuse optimization
results. These signatures are cheap to compute but inexact and hence we
only use this plan cache for more than 10 interesting points which makes
collisions practically impossible.

On a scenario of 1000 iterations over lenet (mini-batch algorithm, with
need to dynamic recompilation, and a large partition of 14 interesting
points in the inner loop), this patch improved the total runtime from
442s to 183s (239s without codegen). This improvement is due to a cache
hit rate of 990/1013 for amenable plans, reducing the number of costed
plans from 11M to 150K (out of a total plan space of 1,064,354,993,445)
and the total codegen overhead from 236s to 8.6s.