[SYSTEMML-2132] Fix intermediate memory estimate matrix multiply hops
This patch fixes a long hidden issue of unnecessarily large intermediate
memory estimate for aggregate binary (i.e., matrix multiply) operators.
This estimate is supposed to account for a potential switch to sparse
(the worst case at the sparsity turn point) but because we used this
turn point, it was already estimated as dense. With this patch, we no
longer compile unnecessary spark instructions for auto-encoder over a
100K x 10K dense input with 35GB driver.