[SYSTEMML-2133] Performance cp maxpooling operations (load balance)
This patch makes a minor performance improvement to maxpooling and
maxpooling_backward operations. So far we used static task partitioning
in the number of threads which led to suboptimal load balance. Since
these maxpooling operations do not create intermediates, we now use
twice as many tasks as threads. On an end-to-end cnn scoring
application, this improved the maxpooling performance by 25%.