[SYSTEMML-2136] Fix int overflow in spark/mr text to binary conversion
This patch fixes an issue of integer overflows when converting matrices
with a large number of rows or columns (>2 billion) from textcell to
binary block representation. The issue was a misplaced cast in the
computation of block sizes.

Furthermore, this also includes a minor javadoc fix of the xor codegen
support (which currently breaks the distribution build).