[SYSTEMML-2041] Performance/memory-efficiency ultra-sparse conv2d
This patch improves the performance and memory-efficiency of
ultra-sparse conv2d operations by allocating the output in sparse
format, forcing ultra-sparse matrix multiplications (with sparse output)
and generalizing all related operations to handle sparse intermediates. 

Furthermore, this also includes a fix for ultra-sparse matrix
multiplications where the rhs is much larger than the lhs which triggers
parallelization over rows in the rhs instead of the lhs, which is not
supported by ultra-sparse matrix multiplications.