[SYSTEMML-2147] Performance read of sparse short-wide binary matrices
This patch improves the performance of reading sparse but short and wide
binary matrices with the multi-threaded reader. This reader synchronized
over the first row per row of blocks which leads to thread contention in
case of threads > num row blocks. We now preallocate all rows in such
scenarios to allow a more fine-grained synchronization without hurting
performance for the common case of tall and skinny matrices.