[SYSTEMML-2152] Fix spark mapmm selection (missing dims constraint)
This patch fixes the operator selection logic for the broadcast-based
mapmm. Since the broadcast matrix is created in the driver, the matrix
needs to satisfy the CP dimensions constraint of int32, which can be
exceeded for ultra-sparse broadcast choices.

In the future, we will also support the creation of broadcast from
blocked RDDs and hdfs files, which will allow to remove the constraint.