[MINOR] Fine-tuning codegen row ops (load balance, nnz recompute)
This patch addresses minor issues of unnecessary many tasks (which is
unnecessary for small inputs) and redundant nnz recomputation in spark
codegen row ops.

Furthermore, this also includes a minor fix of the maintenance of
codegen optimization stats, which did not properly account for the plans
evaluated by the opening heuristic.