[SYSTEMML-2157] Fix parfor optimizer side effect on codegen plans
The parfor optimizer applies a variety of rewrites some which require
the recompilation of the parfor body program or parts of it. Most of
these recompilations leverage the recompiler and thus work properly with
codegen. However, there are two rewrites (removal of branches and the
injection of spark checkpoints) that need to recreate not just
instructions but a partial runtime program. This code path was not
integrated with codegen yet and hence led to lost fusion plans whenever
these rewrites triggered (e.g., for Kmeans over a specific range of data
sizes). This patch fixes the underlying code path for partial program
recompilations, which ensures robustness independent of concrete
instances of rewrites. We also modified the Kmeans tests to check for
properly compiled row-wise templates in the inner loop.