[SYSTEMML-2163] Performance large partitioned spark broadcasts >2GB
This patch makes two minor performance improvements for large, i.e.,
partitioned spark broadcasts. Due to the 2GB limitation for spark
broadcasts, we use partitioned broadcasts, where large side inputs are
chunked up into multiple broadcasts along with an abstraction that hides
this underlying storage from the spark instructions. 

With this patch, (1) we use a more aggressive chunking closer to the 2GB
limit which reduces the number of broadcasts but still ensures that we
never exceed the limit, and (2) we redundantly maintain the meta data in
the wrapper object, which avoids unnecessary broadcast fetches of b[0]
if not all chunks are required by all executors.