[SYSTEMML-2166] Fix scrambled print order after function inlining
With SystemML 1.1 a number of new rewrites were introduced that merge
sequences of statement blocks, which for example appear after branch
removal and function inlining. Although this greatly improves
performance for certain workloads, it might lead to a reordering of
operations without dependencies (such as prints) that end up in the
consolidated DAG. SYSTEMML-2050 tried to address this by ordering of
prints by their line numbers. Unfortunately, this works only in certain
situations because after function lining the print line numbers do not
correspond to the order of the original function calls. 

This patch fixes this issue by merging the non-overlapping roots of
statement blocks s1 and s2 in order of (s1,s2) instead of (s2,s1), where
the latter originated from the fact that we merge s1 into s2.