[SYSTEMML-2130] Fix pack correct sparse/dense format and nnzs, II
This second fix pack on correct sparse/dense representations and nnz
maintenance, fixes additional issue pointed out by #741. In detail, this
includes:

(1) Fix for the nnz maintenance of tsmm outputs in the special case of
zero diagonal elements which is very rare because it requires (except
for cancellations) that entire columns are empty.

(2) Conditional representation change for external UDF outputs, of the
UDFs themselves output matrices in form format.

(3) Rand representation change for distributed spark operations and
conditionally also for CP in-memory operations.

(4) Improved handling of guarded sparsity checks for unary and binary
operations.