[SYSTEMML-2170] Fix parfor optimizer exec type selection (valid dims)
This patch fixes the parfor optimizer rewrite for selecting the
execution mode (local vs remote parfor) to properly account for valid CP
dimensions in addition to the satisfied memory budget constraints.
Previously, we did not check the dimensions, which led to invalid
decisions and thus crashes for ultra-sparse matrices with dimensions
larger than max integer but small memory requirements.