[SYSTEMML-2171] Improved caching of ultra-sparse matrices
This patch improves the checkpoint (i.e., distributed caching) logic for
ultra-sparse matrices. Specifically, we now account for invalid CP
dimensions not just memory to decide upon the compilation of checkpoint
instructions. Furthermore, we also avoid too eager coalesce on
checkpoints if the existing number of partitions is already below the
default parallelism of the cluster.