[SYSTEMML-2174] Performance spark cpmm over ultra-sparse inputs
This patch improves the performance of spark cross-product matrix
multiply (cpmm) operators with single-block outputs to prune empty
blocks before and after the join. Furthermore, we now also avoid
unnecessary deep copies of input blocks.

On an end-to-end scenario with large ultra-sparse matrices and repeated
cpmm executions this reduced the time spent in cpmm from 290s to 135s.