[SYSTEMML-2176] Performance spark reshape ultra-sparse matrices
This patch makes a significant performance improvement to ultra-sparse
distributed spark reshape operations, especially for special cases of
vector-matrix reshapes. In detail, this includes a better enumeration of
output block indexes (which becomes a bottleneck for ultra-sparse
matrices if not done right), avoid unnecessary sort for column-wise
reshape of vectors to matrices, and a guard to ensure correct
sparse/dense output block formats.

On a specific algorithm with 110 distributed reshape operations over
ultra-sparse inputs and intermediates, this patch improved the
end-to-end runtime from 508s to 139s.