[SYSTEMML-2180] Fix codegen row opening condition for right indexing
This patch fixes an issue of invalid generated code for the fuse_all
heuristic with disabled rewrites. Specifically, this fixes the OFMC
opening condition of row templates to only consider indexing operations
with nrow > 1. With rewrites (which is the default) the issue did not
show up because the problematic left indexing was eliminated.