[SYSTEMML-2077] Extended IPA w/ conditional removal of unused functions
So far, inter-procedural analysis (IPA) removed all unused functions,
i.e., functions that are either inlined in all call locations or simply
not reachable from the main problem. This is especially important for
script with nested imports, which otherwise cause unnecessary
compilation and memory overheads. However, this removal is invalid in
the presence of second-order eval functions, which can executed
dynamically constructed function names and hence was temporarily
disabled.

This patch extends IPA to determine if the function call graph contains
at least one call to eval and conditionally disable the removal of
unused functions accordingly. Thus, the common case without eval is not
affected while support eval calls of "unused" functions if a script
requires this flexibility (e.g., for ensemble learning).