[SYSTEMML-2178,2183] Revert ANTLR version upgrade (back to 4.5.3)
This patch reverts the ANTLR version update from 4.5.3 to 4.7.1 for
Spark 2.3 because it lost backwards compatibility with Spark 2.1 and
2.2. Compiling our source with an older version works with warnings on
Spark 2.3 and without issues on Spark 2.1 and 2.2, while a newer version
would produce runtime errors on Spark 2.1 and 2.2. 

In our SystemML 1.1 release notes we should deprecate the support for
Spark 2.1 and 2.2 and in SystemML 1.2 make the switch to the newer ANTLR
version.