[SYSTEMML-2184] Apply basic wdivmm conservatively (sparse/dense inputs)
This patch makes a compiler improvement to apply the fused basic wdivmm
for the pattern (X * U%*%t(V)) more conservatively only for when (1) X
is sparse and U/V are unknown or dense, or (2) when U and V known to be
dense, which are the primary applications for this sparsity exploiting
fused operator. At the same time, this helps to avoid performance issues
for scenarios where U and V are actually ultra-sparse and wdivmm would
hurt performance.

For example, on perftest stratstats 10K x 10K, this patch improved
performance from 128s to 29s end-to-end.