[SYSTEMML-2185] Shuffle tuning spark binary matrix-matrix operations
This patch tunes the shuffle behavior of spark binary matrix-matrix
operations by setting the preferred number of partitions for the join.
This operation has the characteristics that the shuffled intermediate is
larger than the input/output. Hence, we now compute a conservative
preferred number of partitions and explicit enforce the number of
partitions unless the input RDDs carry an existing hash partitioning.

For example, on stratstats 100K x 1K, this patch improved end-to-end
performance from 1100s to 740s.