[SYSTEMML-2188] Fix unnecessary evictions to local fs on rdd collect
This patch improves the performance of large rdd collect operations. For
robustness regarding potential OOMs, we already had functionality for
guarded collects that write the RDD to hdfs and read it into memory
instead of collect because the latter requires twice the memory of a
simple read. However, there are scenarios, where we collect an RDD and
because its size exceeds the buffer pool, we immediately evict it to
local file system in a single-threaded manner.

This patch explicitly considers the buffer pool limit when deciding for
in-memory collect. Furthermore, this also includes some minor
refactorings as a precondition for additional improvements regarding
unnecessary evictions.