[SYSTEMML-2200] Fix unnecessary shuffle in distributed codegen ops
This patch makes a general improvement to distributed codegen operations
on spark. So far, the first input was always assumed to be an RDD and
the inputs to a fused operator where ordered accordingly. However, in
the presence of unknowns, there were scenarios with unnecessary RDD
inputs and thus, unnecessary shuffle due to distributed joins (with
large performance impact). We now decide on broadcasts in a more
generalized way. Furthermore, this also fixes the accounting of
currently pinned broadcasts, which was way too conservative in a parfor
context.

For example, on perftest Kmeans 10Mx1K with 10 concurrent runs, this
patch improved the end-to-end runtime from 9305s to 491s.