[SYSTEMML-2201] Fix removeEmpty empty.return handling in cp/mr
SYSTEMML-2135 introduced the handling of matrices with zero rows and
columns in order to resolve (among others) special cases of removeEmpty,
where we returned a single empty row/column for empty inputs. For
backwards compatibility, we introduced an optimal parameter
'empty.return=true' to explicitly influence this behavior. However, this
parameter was mistakenly ignored in CP and MR instructions.