[SYSTEMML-2207] New optional 'secondorder' arg for external functions
This patch adds an optional argument to the specification of external
functions in order to mark them as 'secondorder', i.e., indicate that
they are calling other functions. Similar to the builtin eval functions,
this flag is in turn used to disable the removal of unused functions,
which might be called within this second order function.