[SYSTEMML-2210] New exists builtin function to check existence of vars
This patch introduces an initial version of the new exists builtin
function that allows to check for the existence of variables by explicit
references to data identifiers or by variable names.

Note that this initial version requires that the referenced variable is
already bound to its logical variable name and that the variable is used
after the existence check (which guarantees its still available as live
variable). In another patch, we will generalize the data dependency
handling to remove these limitation.s